add breadcrumb for navigation in search book, change style of flash messages, add condition for commande deleteUser

+++ src/Services/User/UserService.php
@@ -61,7 +61,7 @@ class UserService {
             $deleted = 0;
             
             foreach($users as $user){
-                if($user->getCreatedAt() > $now->modify('- 14 day') ){
+                if($user->getCreatedAt() < $now->modify('- 14 day') && $user->getAccountValidate() === false){
                     $this->entityManager->remove($user);
                     $this->entityManager->flush();
                     $deleted ++;
